requirements.txt
pyrogram
tgcrypto
aiohttp
requests
asyncio
telegraph
pillow
yt-dlp
python-dateutil
pymongo
aiofiles
python-dotenv
moviepy